+++ OrderApp/View models/SettingsViewModel.swift
@@ -5,7 +5,6 @@
 //  Created by JeongTaek Han on 2023/05/08.
 //
 
-import Foundation
 import UIKit
 
 
@@ -65,111 +64,5 @@ extension MainSettingsViewModel: SettingPresentableDelegate {
             presentAction(item, viewController)
         }
     }
-    
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-final class AppearanceSettingsViewModel: SettingPresentable {
-    
-    private var appearanceSettingController = AppearanceSettingController()
-    
-    var items: [Item] {
-        [
-            Item(
-                title: "System Default",
-                section: .theme,
-                isGroup: false,
-                isChecked: appearanceSettingController.isUnspecifiedColorSchema
-            ),
-            Item(
-                title: "Light Mode",
-                section: .theme,
-                isGroup: false,
-                isChecked: appearanceSettingController.isUnspecifiedColorSchema
-            ),
-            Item(
-                title: "Dark Mode",
-                section: .theme,
-                isGroup: false,
-                isChecked: appearanceSettingController.isUnspecifiedColorSchema
-            ),
-        ]
-    }
-    
-    enum Section: SettingSectionPresentable {
-        case theme
-        
-        var title: String? {
-            switch self {
-            case .theme:
-                return "Theme"
-            }
-        }
-        
-        var description: String? {
-            switch self {
-            case .theme:
-                return "Configure app's colot theme schema."
-            }
-        }
-    }
-    
-    struct Item: SettingItemPresentable {
-        
-        var title: String
-        var description: String?
-        var section: Section
-        
-        var isGroup: Bool
-        var isChecked: Bool
-        
-    }
-    
-}
-
-protocol SettingSectionPresentable: Hashable {
-    
-    var title: String? { get }
-    var description: String? { get }
-    
-}
-
-protocol SettingItemPresentable: Hashable {
-    
-    associatedtype Section: SettingSectionPresentable
-    
-    var title: String { get }
-    var description: String? { get }
-    var section: Section { get }
-    
-    var isGroup: Bool { get }
-    var isChecked: Bool { get }
-    
-}
-
-protocol SettingPresentable {
-    
-    associatedtype Section: SettingSectionPresentable
-    associatedtype Item: SettingItemPresentable
-    
-    var items: [Item] { get }
-    
-}
-
-
-protocol SettingPresentableDelegate: AnyObject {
-    
-    func provideSettingViewController(of item: any SettingItemPresentable, presentAction: (any SettingItemPresentable, UIViewController) -> Void)
     
 }
